+++ packages/protocol-kit/src/Safe.ts
@@ -11,13 +11,6 @@ import {
   TransactionOptions,
   TransactionResult
 } from '@safe-global/safe-core-sdk-types'
-import {
-  SAFE_FEATURES,
-  hasSafeFeature,
-  isMetaTransactionArray,
-  isSafeMultisigTransactionResponse,
-  sameString
-} from './utils'
 import ContractManager from './managers/contractManager'
 import FallbackHandlerManager from './managers/fallbackHandlerManager'
 import GuardManager from './managers/guardManager'
@@ -25,11 +18,18 @@ import ModuleManager from './managers/moduleManager'
 import OwnerManager from './managers/ownerManager'
 import { ContractNetworksConfig } from './types'
 import {
+  EthSignSignature,
+  hasSafeFeature,
+  isMetaTransactionArray,
+  isSafeMultisigTransactionResponse,
+  SAFE_FEATURES,
+  sameString
+} from './utils'
+import {
   generateEIP712Signature,
   generatePreValidatedSignature,
   generateSignature
-} from './utils/signatures'
-import SafeSignature from './utils/signatures/SafeSignature'
+} from './utils/signatures/utils'
 import EthSafeTransaction from './utils/transactions/SafeTransaction'
 import { SafeTransactionOptionalProps } from './utils/transactions/types'
 import {
@@ -800,7 +800,7 @@ class Safe {
     }
     const safeTransaction = await this.createTransaction({ safeTransactionData })
     serviceTransactionResponse.confirmations?.map((confirmation) => {
-      const signature = new SafeSignature(confirmation.owner, confirmation.signature)
+      const signature = new EthSignSignature(confirmation.owner, confirmation.signature)
       safeTransaction.addSignature(signature)
     })
     return safeTransaction
