+++ packages/protocol-kit/src/utils/index.ts
@@ -1,5 +1,6 @@
+export * from './address'
 export * from './eip-3770'
 export * from './eip-712'
 export * from './safeVersions'
-export * from './address'
+export * from './signatures'
 export * from './transactions'
